requirements.txt
about-time==3.1.1
aiohttp==3.8.1
aiosignal==1.2.0
alive-progress==2.4.1
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
discord-sentry-reporting==2.0.4
discord.py==2.0.1
discord==2.0.0
dpytest==0.5.3
Flask==2.2.2
frozenlist==1.3.1
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.12.4
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grapheme==0.6.0
grpc-google-iam-v1==0.12.4
grpcio-status==1.48.1
grpcio==1.48.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
oauth2client==4.1.3
oauthlib==3.2.1
peewee==3.15.2
proto-plus==1.22.1
protobuf==4.21.5
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pygit2==1.10.1
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
pytest
pyparsing==3.0.9
python-dotenv==0.21.0
requests-oauthlib==1.3.1
requests==2.28.1
rsa==4.9
sentry-sdk==1.9.8
setuptools==60.2.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
